Move IBitmapFactory to WPF library (should probably have been there to start with). Pass the transformation  matrix to CreateBitmap method Still needs a little work as the timing isn't quite right.
