+++ CefSharp.Wpf/Rendering/BitmapFactory.cs
@@ -2,18 +2,15 @@
 //
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
+using System.Windows.Media;
 using CefSharp.Internals;
-using System.Windows;
 
 namespace CefSharp.Wpf.Rendering
 {
     public class BitmapFactory : IBitmapFactory
     {
-        public BitmapInfo CreateBitmap(bool isPopup)
+        public BitmapInfo CreateBitmap(bool isPopup, Matrix matrix)
         {
-            var presentationSource = PresentationSource.FromVisual(Application.Current.MainWindow);
-            var matrix = presentationSource.CompositionTarget.TransformToDevice;
-
             if (matrix.M11 > 1.0 || matrix.M22 > 1.0)
             {
                 return new WritableBitmapInfo(96*matrix.M11, 96*matrix.M22)
