+++ CefSharp.Wpf/Rendering/IBitmapFactory.cs
@@ -2,12 +2,13 @@
 //
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
 
+using System.Windows.Media;
 using CefSharp.Internals;
 
-namespace CefSharp
+namespace CefSharp.Wpf.Rendering
 {
     public interface IBitmapFactory
     {
-        BitmapInfo CreateBitmap(bool isPopup);
+        BitmapInfo CreateBitmap(bool isPopup, Matrix matrix);
     }
 }
